Conflicts:
        clang/include/clang/Basic/LangOptions.def
        clang/include/clang/Driver/CC1Options.td
        clang/lib/Basic/IdentifierTable.cpp
        clang/lib/Frontend/CompilerInvocation.cpp
        clang/lib/Frontend/InitPreprocessor.cpp
        clang/lib/Headers/CMakeLists.txt
        clang/lib/Sema/SemaStmtAsm.cpp
        clang/lib/Sema/SemaType.cpp

Other manually modified files:
        buildbot/configure.sh
        clang/lib/AST/ASTContext.cpp
        clang/lib/AST/ItaniumMangle.cpp
        clang/lib/CodeGen/BackendUtil.cpp
        clang/lib/CodeGen/CodeGenFunction.cpp
        clang/lib/CodeGen/CodeGenModule.cpp
        clang/lib/CodeGen/CodeGenTypes.cpp
        clang/lib/CodeGen/ModuleBuilder.cpp
        clang/lib/Frontend/CompilerInstance.cpp
        clang/lib/Parse/ParseAST.cpp
        clang/lib/Sema/Sema.cpp
        clang/lib/Sema/SemaTemplateInstantiateDecl.cpp

+++ clang-tools-extra/clangd/Diagnostics.cpp
@@ -161,11 +161,11 @@ std::string capitalize(std::string Message) {
 ///
 ///     dir1/dir2/dir3/../../dir4/header.h:12:23
 ///     note: candidate function not viable: requires 3 arguments
-std::string mainMessage(const Diag &D) {
+std::string mainMessage(const Diag &D, bool DisplayFixesCount) {
   std::string Result;
   llvm::raw_string_ostream OS(Result);
   OS << D.Message;
-  if (!D.Fixes.empty())
+  if (DisplayFixesCount && !D.Fixes.empty())
     OS << " (" << (D.Fixes.size() > 1 ? "fixes" : "fix") << " available)";
   for (auto &Note : D.Notes) {
     OS << "\n\n";
@@ -252,7 +252,7 @@ void toLSPDiags(
 
   {
     clangd::Diagnostic Main = FillBasicFields(D);
-    Main.message = mainMessage(D);
+    Main.message = mainMessage(D, Opts.DisplayFixesCount);
     if (Opts.EmbedFixesInDiagnostics) {
       Main.codeActions.emplace();
       for (const auto &Fix : D.Fixes)
@@ -335,6 +335,11 @@ void StoreDiags::HandleDiagnostic(DiagnosticsEngine::Level DiagLevel,
 
     llvm::SmallVector<TextEdit, 1> Edits;
     for (auto &FixIt : Info.getFixItHints()) {
+      // Follow clang's behavior, don't apply FixIt to the code in macros,
+      // we are less certain it is the right fix.
+      if (FixIt.RemoveRange.getBegin().isMacroID() ||
+          FixIt.RemoveRange.getEnd().isMacroID())
+        return false;
       if (!isInsideMainFile(FixIt.RemoveRange.getBegin(),
                             Info.getSourceManager()))
         return false;
